Tectonic Explorer: move all CSS modules to components directory so they're next to related TSX file
[#184314710]
