+++ packages/tectonic-explorer/js/components/about-dialog.scss
