+++ packages/tectonic-explorer/js/components/data-collection-dialog.scss
