+++ packages/tectonic-explorer/js/components/exit-data-collection-dialog.scss
