+++ packages/tectonic-explorer/js/components/icon-highlight-button.scss
@@ -1,4 +1,4 @@
-@import "../css/variables.scss";
+@import "./variables.scss";
 
 .iconButton {
   display: flex;
