+++ packages/tectonic-explorer/js/components/keys/rock-images.scss
