+++ packages/tectonic-explorer/js/components/keys/rock-types.scss
@@ -1,4 +1,4 @@
-@import "../../css/variables.scss";
+@import "../variables.scss";
 
 .rockKey {
   position: relative;
