+++ packages/tectonic-explorer/js/components/keys/seismic-data.tsx
@@ -4,7 +4,7 @@ import { EARTHQUAKE_COLORS } from "../../plates-view/earthquake-helpers";
 import { BaseComponent, IBaseProps } from "../base";
 import config from "../../config";
 
-import css from "../../../css-modules/keys/seismic-data.scss";
+import css from "./seismic-data.scss";
 
 interface IState { }
