+++ packages/tectonic-explorer/js/components/map-type-button.scss
@@ -1,4 +1,4 @@
-@import "../css/variables.scss";
+@import "./variables.scss";
 
 .mapTypeButton {
   display: flex;
