+++ packages/tectonic-explorer/js/components/mode-wizard.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Button } from "react-toolbox/lib/button";
 import EarthImage from "../../images/earth@3x.png";
 
-import modeWizardTheme from "../../css-modules/mode-wizard-theme.scss";
+import modeWizardTheme from "./mode-wizard-theme.scss";
 import "../../css/mode-wizard.scss";
 
 interface IModeConfig {
