+++ packages/tectonic-explorer/js/components/side-container.scss
@@ -1,4 +1,4 @@
-@import "../css/variables.scss";
+@import "./variables.scss";
 
 // General styling of tabs. If it ever gets extracted, these styles should go to common component.
 $outlineColor: #797979;
