+++ packages/tectonic-explorer/js/components/slider-switch.scss
@@ -1,4 +1,4 @@
-@import "../css/variables.scss";
+@import "./variables.scss";
 
 .sliderSwitch {
   position: relative;
