+++ packages/tectonic-explorer/js/components/top-bar.scss
@@ -1,4 +1,4 @@
-@import "../css/variables.scss";
+@import "./variables.scss";
 
 .topBar {
   background: #fff;
