+++ packages/tectonic-explorer/js/components/variables.scss
@@ -1,0 +1,24 @@
+$primaryColor: #F8C84E;
+$inactiveColor: #9e9e9e;
+$topBarHeight: 20px;
+$default-font-family: "Lato", verdana, arial, helvetica, sans-serif;
+$default-font-size: 14px;
+$smaller-font-size: 13px;
+$default-font-weight: bold;
+$component-border-radius: 9px;
+$control-panel-height: 65px;
+$control-panel-content-height: 74px; // it's higher than $control-panel-height because of "bubbles"
+$tabs-panel-width: 336px;
+
+// PostCSS variables
+:root {
+  --theme-primary-color: 255, 255, 255;
+  --theme-secondary-color: 51, 51, 51;
+  --theme-primary-variant-color: 245, 245, 245;
+  --theme-selected-background-color: 234, 234, 234;
+  --theme-selected-text-color: 51, 51, 51;
+  --theme-hover-background-color: 225, 225, 225;
+  --theme-hover-text-color: 51, 51, 51;
+  --theme-active-background-color: 225, 225, 225;
+  --theme-active-text-color: 51, 51, 51;
+}
